feat: render markdown as html (#10)
* feat: add render subcommand

* feat: grab the file content from a path and render it as html

* refactor: move markdown impl into lib

* refactor: move markdown impl into domain

* refactor: move dice impl into domain

* add render just command
