+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pinbreak"
 version = "0.1.0-alpha"
-description = "CLI tool to roll dice (for now)"
+description = "CLI tool to roll dice and render markdown files (for now)"
 edition = "2024"
 readme = "docs/README.md"
 repository = "https://github.com/rrvsh/pinbreak"
@@ -11,6 +11,7 @@ categories = ["command-line-utilities"]
 
 [dependencies]
 clap = { version = "4.5.53", features = ["derive"] }
+markdown = "1.0.0"
 rand = "0.9.2"
 
 [lints.clippy]
